containers/seed-data/pyproject.toml
[project]
name = "private-vectordb-seed-data"
version = "0.1.0"
description = "Seeds data into the datastore of the private vectordb"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.54",
    "lancedb>=0.15.0",
    "pandas>=2.2.3",
    "pylance>=0.19.1",
    "python-dotenv>=1.0.1",
]
